requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
pytest==6.2.5
Flask-SQLAlchemy==2.5.1